Reorganized file setup
