+++ app/src/main/res/drawable/ic_home.xml
@@ -7,7 +7,7 @@
   <group android:scaleX="0.58"
       android:scaleY="0.58"
       android:translateX="5.04"
-      android:translateY="10.04">
+      android:translateY="5.04">
     <path
         android:fillColor="@android:color/white"
         android:pathData="M10,20v-6h4v6h5v-8h3L12,3 2,12h3v8z"/>
